feat: refactor ingress configuration and rename hostapi to experienceapi
- Moved ingress rules from Tiltfile to Helm chart auto-generation
- Renamed hostapi.py to experienceapi.py with /experience prefix
- Removed root_path configuration and simplified routing
- Updated Helm chart to provide better defaults for ingress
- Improved separation of concerns between Tiltfile and Helm chart
