+++ Tiltfile
@@ -16,6 +16,20 @@ app_env = os.getenv('APP_ENV', 'dev')
 load('ext://helm_resource', 'helm_resource', 'helm_repo')
 helm_repo('dev-util', 'https://whale-net.github.io/dev-util')
 
+# setup nginx ingress controller for local development
+helm_repo('ingress-nginx', 'https://kubernetes.github.io/ingress-nginx')
+helm_resource('manman-nginx-ingress', 'ingress-nginx/ingress-nginx',
+    namespace='ingress-nginx',
+    flags=['--create-namespace',
+           '--set=controller.service.type=NodePort',
+           '--set=controller.hostPort.enabled=true',
+           '--set=controller.service.nodePorts.http=30080',
+           '--set=controller.service.nodePorts.https=30443',
+           '--set=controller.admissionWebhooks.enabled=false',
+           '--set=controller.ingressClassResource.name=manman-nginx',
+           '--set=controller.ingressClass=manman-nginx']
+)
+
 # setup postgres
 helm_resource('postgres-dev', 'dev-util/postgres-dev', resource_deps=['dev-util'],
     flags=['--set=postgresDB=manman', '--set=namespace={}'.format(namespace)]
@@ -61,7 +75,10 @@ helm_set_args = [
     'migrations.skip_migration=true',
     # Control which APIs to deploy
     'apis.experience.enabled={}'.format(str(enable_experience_api).lower()),
-    'apis.workerDal.enabled={}'.format(str(enable_worker_dal_api).lower())
+    'apis.workerDal.enabled={}'.format(str(enable_worker_dal_api).lower()),
+    # Enable ingress for local development with auto-generated rules
+    'ingress.enabled=true',
+    'ingress.ingressClassName=manman-nginx',
 ]
 
 k8s_yaml(
@@ -73,11 +90,16 @@ k8s_yaml(
     )
 )
 
-# Port forward for Experience API (user-facing)
-k8s_resource(workload='manman-experience-{}'.format(app_env), port_forwards='8000:8000')
+# Note: Ingress controller is available on NodePort 30080
+# APIs are auto-routed by Helm chart ingress template:
+#   Experience API: http://localhost:30080/experience/
+#   Worker DAL API: http://localhost:30080/workerdal/
 
-# Port forward for Worker DAL API (internal API for workers)
-k8s_resource(workload='manman-worker-dal-{}'.format(app_env), port_forwards='8001:8000')
+print("APIs will be available at (auto-generated by Helm chart):")
+if enable_experience_api:
+    print("  Experience API: http://localhost:30080/experience/")
+if enable_worker_dal_api:
+    print("  Worker DAL API: http://localhost:30080/workerdal/")
 
 # this should be a docker compose becauset hat is how I will actually deploy it
 #local_resource(
