+++ charts/manman-host/values.yaml
@@ -22,6 +22,44 @@ apis:
     port: 8000
     command: "start-worker-dal-api"
 
+# Ingress configuration (primarily for development)
+ingress:
+  # Enable/disable ingress
+  enabled: false
+
+  # Custom ingress name (defaults to manman-ingress-{app_env} if empty)
+  name: ""
+
+  # Custom host (defaults to localhost for development if empty)
+  host: "localhost"
+
+  # Ingress class name (e.g., nginx, traefik)
+  ingressClassName: ""
+
+  # Custom annotations for ingress controller
+  annotations: {}
+    # nginx.ingress.kubernetes.io/rewrite-target: /$2
+    # nginx.ingress.kubernetes.io/use-regex: "true"
+    # cert-manager.io/cluster-issuer: letsencrypt-prod
+
+  # Override auto-generated rules with explicit configuration
+  # If empty, rules are auto-generated based on enabled APIs
+  # Auto-generated rules will create:
+  #   - /experience -> experience API service
+  #   - /workerdal -> worker DAL API service
+  rules: []
+    # Example explicit rules:
+    # - host: localhost
+    #   paths:
+    #     - path: /experience
+    #       pathType: Prefix
+    #       serviceName: manman-experience-dev-service
+    #       servicePort: 8000
+    #     - path: /workerdal
+    #       pathType: Prefix
+    #       serviceName: manman-worker-dal-dev-service
+    #       servicePort: 8000
+
 # Migration configuration
 migrations:
   # Set to true to skip running migrations entirely
